fix: update package.json version and restructure main entry point; remove TypeScript configuration and migrate to JavaScript
